requirements.txt
# Core dependencies
streamlit>=1.0.0
geopy>=2.3.0
reportlab>=3.6.12
pandas>=1.3.0
python-dotenv>=0.19.0
requests>=2.26.0
numpy>=1.21.0
openpyxl>=3.0.7  # For Excel file support

# Testing dependencies
pytest>=7.0.0
pytest-cov>=3.0.0  # For coverage reporting
pytest-mock>=3.6.1  # For mocking
requests-mock>=1.9.3  # For mocking HTTP requests

# Python version
python>=3.11
